backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pyyaml
httpx
openai
